requirements.txt
asgiref==3.2.3
Babel==2.8.0
certifi==2019.11.28
Django==2.2.8
django-jsonfield==1.4.0
django-notify-x==0.1.9
django-phonenumber-field==4.0.0
djangorestframework==3.11.0
phonenumbers==8.11.1
pytz==2019.3
six==1.13.0
sqlparse==0.3.0

